Add files via upload
The purpose of this project is to create a digital magazine 
cover displaying a public figure.
 The assets that i used for this project are actual non fictional photos of my mom 
that were featured many years ago during her job as a Bear Biologist for the 
state of NJ Division of Fish Game & Wildlife. You can feel free to google her.
 My mom was often featured in magazines and newspapers and traveled to 
various places to give talks.
 I used the transform tool to rotate the magazine cover to a vertical position.
 I used the scale tool to place the magazine page appropriately on the canvas.
 I imported the smaller thumbnail size photos of my mom by copying and 
pasting them onto the magazine cover.
 I used the text tool to write a date for the magazine as well as a name for
 the magazine.
 I used the healing brush tool cover up some original information on 
the photos and to make my mom"s hair not so gray. 
I used the brackets on my keyboard to adjust my healing brush.
 I used the magnifying glass on the left side of the screen to zoom in and
 out of the magazine cover as I worked on it.
 I used adjustment layers to adjust the brightness of the photos.
 I used the neural filter to colorize the photos.
 I used Photoshop express to generate a mock qr code that I copied & pasted 
onto the digital magazine cover in Photoshop.
 It was a bit of a challenge to work with the text tool but I feel confident 
that I am on my way to becoming a pro at using the text tool.